Version 1.3 DeltaImport and Bugfixes, Working Implementation
The state of the code as of this commit works properly, and has been
tested to work with DynamoDB. The ImportHandler is high performance as
well.
Some work still needs to be done, the ON-ERROR behavior around certain
exceptions may not be correct, for example for explicit type mapping,
errors are probably not handled correctly, or the ON-ERROR value is not
respected.
I added quite a lot of useful debugging in this work.  If you name your
table improperly a list of tables will be returned (if you have
permission) If you don't configure your ValueMap or NameMap correctly,
the Table description will be returned along with debugging output of
the resultant query you submitted.  I also check for empty values in the
ValueMap, for example if you use a solr configuration variable, but the
variable can't be resolved, a Warning will be logged if the ValueMap
value that is parsed out of your config is empty (the same thing should
be done to NameMap) probably, and other fields - warnings for any field
that had a variable in it, but is empty after variable expansion.  The
DELTA IMPORT works well, and I added support for 2 custom variables that
provide you with epoch-seconds and epoch-milliseconds for your delta
query... I will add an example of how to use them, along with conditions
in the data import handler.  Currently the DELTA IMPORT does NOT support
deleting records, this is supported by the framework and I will add this
functionality in the future.  A full-scan will be needed to perform the
set-comparison in order to delete records.